feat(behavior): add edge-tooltip
